update certificate creation section
Fixed a few things and made it clear that there are two approaches to the certificate creation process.
I was about to change the LM\Root store to LM\My but you already did that. I was wondering why it was there in the first place. heh :-)

Let me know what you think.